+++ src/fireworks/training/cookbook/utils/__init__.py
@@ -1,20 +1,23 @@
-"""Cookbook utilities -- infrastructure, losses, data, logging, and more."""
+"""Cookbook utilities -- infrastructure, losses, data, logging, and more.
+
+RL-specific utilities (GRPO/DAPO/GSPO losses, TIS, R3, batching) live in
+``fireworks.training.cookbook.utils.rl``.
+"""
 
 __all__ = [
     "DEFAULT_ADAM",
     "DeployConfig",
     "EvalFn",
     "HotloadConfig",
-    "ISConfig",
     "InfraConfig",
     "ReconnectableClient",
     "ResumeConfig",
     "RewardFn",
     "StepCallback",
     "WandBConfig",
-    "build_r3_routing_matrices",
     "compute_advantages",
     "create_trainer_job",
+    "resolve_and_apply_shape",
     "encode_text",
     "extract_text",
     "find_common_prefix_length",
@@ -22,8 +25,7 @@ __all__ = [
     "load_preference_dataset",
     "log_metrics_json",
     "make_dpo_loss_fn",
-    "make_grpo_loss_fn",
-    "make_grpo_tis_loss_fn",
+    "make_batch_sft_loss_fn",
     "make_sft_loss_fn",
     "setup_deployment",
     "setup_resume",
@@ -47,6 +49,7 @@ from fireworks.training.cookbook.utils.infra import (
     setup_deployment,
     create_trainer_job,
     setup_training_client,
+    resolve_and_apply_shape,
 )
 from fireworks.training.cookbook.utils.client import ReconnectableClient
 from fireworks.training.cookbook.utils.config import (
@@ -63,7 +66,7 @@ from fireworks.training.cookbook.utils.config import (
 from fireworks.training.cookbook.utils.losses import (
     make_dpo_loss_fn,
     make_sft_loss_fn,
-    make_grpo_loss_fn,
+    make_batch_sft_loss_fn,
 )
 from fireworks.training.cookbook.utils.resume import setup_resume
 from fireworks.training.cookbook.utils.logging import (
@@ -73,5 +76,3 @@ from fireworks.training.cookbook.utils.logging import (
     log_metrics_json,
 )
 from fireworks.training.cookbook.utils.validation import validate_config, validate_preflight
-from fireworks.training.cookbook.utils.router_replay import build_r3_routing_matrices
-from fireworks.training.cookbook.utils.importance_sampling import ISConfig, make_grpo_tis_loss_fn